% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_cup_24_elite_men_simulated}
\alias{world_cup_24_elite_men_simulated}
\title{Elite Men World Cup 2024 Simulated Results}
\format{
A data frame with variables:
\describe{
\item{name}{Rider's name (e.g., "NORTON Dakotah").}
\item{event_name}{Name of the event (e.g., "Fort William").}
\item{split_1}{Rider's fastest first split time (in seconds).}
\item{split_2}{Rider's fastest second split time (in seconds).}
\item{split_3}{Rider's fastest third split time (in seconds).}
\item{split_4}{Rider's fastest fourth split time (in seconds).}
\item{time}{Rider's fastest possible time to complete the course (in seconds).}
\item{rank}{The finishing rank of the rider based on time.}
\item{time_from_leader}{Time difference from the fastest rider (in seconds).}
}
}
\source{
Simulated from section analysis of 2024 World Cup events
}
\usage{
world_cup_24_elite_men_simulated
}
\description{
This dataset takes the fastest section times available (from timed training,
qualifying, semi-finals, and finals) for each rider across at each event and
combines them together to simulate what could have been their fastest race
run. These fastest runs are then ranked to determine a new simulated set of
race results.
}
\keyword{datasets}
